wip: refactor ClaimPlant into class
